requirements.txt
nltk
spacy
python-Levenshtein
